Commonize some items in EOB providers, move tests for commonized methods, update calls, split test that had too many parts, add tests for bad type to EOB tests
